created two separated components for AccountView - Private and Public - so the public one can be accessible for other users from the Chat
