+++ frontend/src/router/index.ts
@@ -15,7 +15,7 @@ const router = createRouter({
       // route level code-splitting
       // this generates a separate chunk (About.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
-      component: () => import("../views/AccountView.vue"),
+      component: () => import("../views/HomeView.vue"),
     },
     {
       path: "/chat",
@@ -47,7 +47,7 @@ const router = createRouter({
       // route level code-splitting
       // this generates a separate chunk (About.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
-      component: () => import("../views/AccountParams.vue"),
+      component: () => import("../views/AccountView.vue"),
     },    
     {
       path: "/user/:id",
